Enforce PWR071 in example codes

+++ Checks/PWD011/example-omp.f90
@@ -1,10 +1,11 @@
 ! PWD011: Missing OpenMP lastprivate clause
 
 real function example(A, B, C)
+  use iso_fortran_env, only: real32
   implicit none
-  real, intent(in) :: A(:), B(:)
-  real, intent(inout) :: C(:)
-  real :: liveOut
+  real(kind=real32), intent(in) :: A(:), B(:)
+  real(kind=real32), intent(inout) :: C(:)
+  real(kind=real32) :: liveOut
   integer :: i
 
   !$omp parallel do private(i, liveOut) shared(A, B, C)
